Complete Add Trader Logic

+++ angular-frontend/angular-trading-ui/src/app/trader-list/trader-list.component.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
 import { TraderListService } from '../trader-list.service';
+import { DialogService } from '../dialog.service';
 
 @Component({
   selector: 'app-trader-list',
   templateUrl: './trader-list.component.html',
   styleUrls: ['./trader-list.component.sass'],
-  providers: [TraderListService],
+  providers: [TraderListService, DialogService],
 })
 export class TraderListComponent {
   constructor(private _traderList: TraderListService) {}
@@ -14,7 +15,8 @@ export class TraderListComponent {
   dataSource = this._traderList.getDataSource();
 
   deleteTrader(id: number): void {
-    if (this._traderList.deleteTrader(id))
-      this.dataSource = this._traderList.getDataSource();
+    if (DialogService.confirm('Delete this trader?'))
+      if (this._traderList.deleteTrader(id))
+        this.dataSource = this._traderList.getDataSource();
   }
 }
